requirements.txt
backoff==2.2.1
google-auth==2.6.2
gspread==5.2.0
requests==2.32.0
discord_webhook==0.15.0
pymongo==4.6.3
pymongo[srv]
python-dotenv==0.20.0
urllib3==1.23
